+++ src/app/dashboard/page.tsx
@@ -1,0 +1,8 @@
+export default function Dashboard() {
+    return (
+      <main>
+        <h1>Dashboard</h1>
+      </main>
+    );
+  }
+
